Améliore le style des références et ajoute un feedback sur la page résultat
